move list cmds into own file
